"fix ci"

+++ python/paddle/fluid/tests/unittests/test_mul_op.py
@@ -83,7 +83,7 @@ class TestFP16MulOp1(OpTest):
         self.use_mkldnn = False
         x = np.random.random((32, 84)).astype("float16")
         y = np.random.random((84, 100)).astype("float16")
-        self.inputs = {'X': x.view(np.uint16), 'Y': y.view(np.uint16)}
+        self.inputs = {'X': x.view(np.float16), 'Y': y.view(np.float16)}
         self.attrs = {'use_mkldnn': self.use_mkldnn}
         self.outputs = {'Out': np.dot(x, y)}
 
@@ -100,7 +100,7 @@ class TestFP16MulOp2(OpTest):
         self.use_mkldnn = False
         x = np.random.random((15, 4, 12, 10)).astype("float16")
         y = np.random.random((4, 30, 8, 2, 9)).astype("float16")
-        self.inputs = {'X': x.view(np.uint16), 'Y': y.view(np.uint16)}
+        self.inputs = {'X': x.view(np.float16), 'Y': y.view(np.float16)}
         self.attrs = {
             'x_num_col_dims': 2,
             'y_num_col_dims': 2,
